MB-63977: Refactor DcpConsumer control negotiation
The aim of this commit is to change DcpConsumer control negotiation so
that for future controls requires only a few lines of change and no
more copy/paste (excluding test code).

The copy/paste "Blocking" and bespoke "handle" functions are all
replaced by a list of controls (pendingControls) with one common send
and receive function to process control reqyest/response.

This brings all DcpConsumer controls under a "blocking" pattern, meaning
each control must wait until a response to the previous. This does
change how some of the controls were previously processed where there
was no wait for a response, but the consistency and structure of the
code is more beneficial here (noting that there was no guarantee that
controls were ever requested in bulk).

Change-Id: Ib2fa557f979960c930ae56794cbcaa07bce51d82
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218861
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>