MB-64071: Add client SDK count to prometheus
Add a metric to Prometheus to track the number of connections per SDK
version up to a maximum of 100 different versions. Once the
maximum is reached the least recently SDK is evicted to start
tracking the newly connected SDK

These numbers are pushed to prometheus by using the key
current_external_client_connections and a label of "sdk" containing
the SDK version.

When using mcstat these stats are reported as:

    current_sdk_connections:<identifier>

When using mcstat you might see:

    current_sdk_connections:mcstat/7.6.4

NOTE: The server only keeps track of the first 32 characters
of the agent name, which means that clients trying to report
an identifier/version which exceeds that length will fail to
parse and end up being reported as "unknown"

Change-Id: I054b1d11588617286355873a6029bc892afc6c32
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218939
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>