[unreviewed-changes] Include " or cc:{user}" in query
Some teams (such as ns_server) use 'CC' instead of 'Reviewer' for
adding reviewers to the change. To support these teams using this
script, it should include the changes that are have at least two of the
team cc'd (or owner/reviewer, as currently).

This change simply adds " or cc:{user}" to each reviewer subquery.

The change was tested for both ns_server and kv_engine, and it provides
the correct changes for both.

Change-Id: I15ddc667fb3700812b968c30b9414d8d4dd23aa5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218988
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>