MB-27980: Add document read/write bytes to slow operation
The request was to include the document size as part of the
slow operation logging. This patch adds something "similar" to
that: the number of bytes we've fetched and written to the engine.

In the case where we perform a single fetch of a document it would
be the same as the document size, but in when using subdoc
this would be the aggregate of all attempts to perform a subdoc
operation (in the case of a race condition causing the server
to retry the entire subdoc transformation).

I believe these numbers are probably more interesting than a
single "document size" number as if there was a heavy race
condition during subdoc one would have reported 10 bytes as
the document size, while the retries could have fetched
documents up to 20MB every time before the final time where
a small document was returned.

Change-Id: I5736d9d3a6619593fc92804e9681f1c8c5d04ef7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219045
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dan Owen <owend@couchbase.com>