[BP] MB-63769: Mark delete created by deleteWithMeta as resident
Previously, we have observed that a deleteWithMeta operation can create
a non-temp. non-resident, deleted item. that is stored in the HT. If a
GET with the flag GET_DELETED_VALUE is called on the same key before
the item has been persisted to disk, then the value will be fetched from
disk but does not exist. Throwing an exception.

In such a case, we should not require a BGFetch and instead return the
delete that is found in the HT because this is always the most recent
value. By setting the resident bit for this item, we avoid queueing a
BGFetch and instead return the item found in the HT.

cherry picked from commit ac5e7ada249d9a7f2e9fdfe81981386146fb51ce 
for 7.6.5-MP3

Change-Id: Ibee887ec625dd80a390acf75b2539ed97905bc7c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223785
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>