MB-63769: Mark delete created by deleteWithMeta as resident
Previously, we have observed that a deleteWithMeta operation can create
a non-temp. non-resident, deleted item. that is stored in the HT. If a
GET with the flag GET_DELETED_VALUE is called on the same key before
the item has been persisted to disk, then the value will be fetched from
disk but does not exist. Throwing an exception.

In such a case, we should not require a BGFetch and instead return the
delete that is found in the HT because this is always the most recent
value. By setting the resident bit for this item, we avoid queueing a
BGFetch and instead return the item found in the HT.

Change-Id: I60d7684741acae9661954a624160b77e7f11e5c8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217866
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>