[dbfill] use atomics instead of locking shared data
Use a atomic to let the producer signal to the consumers that
it is done producing data, and another atomics to let the
consumers signal to the producer that they're idle.

This allows for fixing the problem where some connections
became idle and never sent any data to the server as the
queue was empty when the consumer tried to fill the queue.

Change-Id: Ifbf4f234b5682880681ef7ec99afb642ca0d4588
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219278
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>