MB-64160: Remove the _bytes suffix from kv_ep_mem_*_wat_bytes
The old names of the watermarks metrics did not have _bytes at the end.

I've removed the suffix by setting the unit as none in the
stats_definitions.json. Currently, there does not seem to be another way
to achieve this other than changing the generation code.

This change does mean the metric will be documented as unitless, but
that is not a regression since before the _bytes was added, the metadata
for the metric was generated from configuration.json and thus it was
documented as unitless anyways.

Change-Id: I84cdf1eda25812cad25544826efbe4f9e072efce
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219324
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>