MB-62703: Add kill switch for backfill idle checks
Add a configuration which can disable the DCP idle checks.

For ephemeral buckets the feature is for now disabled, but has test
coverage. The focus is on persistent buckets.

To default the kill switch to false for ephemeral requires a change
to our configuration.json, the addition of a "conditional" object
which allows for the generated config to express.

if (bucket_type == ephemeral) {
  dcp_backfill_idle_protection_enabled = false
} else {
  dcp_backfill_idle_protection_enabled = true
}

The JSON spec for this is to define the "default" value as:

{
  "if": {
    "parameter": {
      "equals": "value",
      "then": "use value1",
      "elif": [
        {
          "parameter" : {
            "equals": "value",
            "then": "use value2"
          }
        }
      ]
    },
    "else": "use value3"
  }
}

Only parameter == value is supported and usage of elif occurs
in the following commit where magma support is disabled.

Change-Id: If6f999310839f69f9844bb4de37072592efeca7d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219639
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>