MB-62703: End a stalled backfill if disk space is low
Add logic to DCP backfill so that more than just a time check
decide if a stream should end. The commit adds file system
usage checks and an additional check to see if closing the scan
could free disk usage.

For example if disk space were critical, and the scan is stalled
but has open the curerent revision of a couchstore file, closing
the snapshot won't improve the disk space situation. But if
compaction ran, the vbucket could now be using 2x disk space and
a scan could now be closed to free up disk space.

This only works for couchstrore scans until we get MB-64220, so
magma is disabled.

configuration.json now uses the "elif" to generate the following

if (bucket_type == ephemeral) {
  dcp_backfill_idle_protection_enabled = false
} else if (backend == magma)
  dcp_backfill_idle_protection_enabled = false
} else {
  dcp_backfill_idle_protection_enabled = true
}

Change-Id: I592ceb464fbf9c5a744bf669ca52188fcb3b2f8f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219699
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>