MB-64205: Default all metrics to "gauge" instead of "untyped"
"untyped" is a safe default to use for Prometheus when the type is not
known (as it is with our metrics since they were migrated from STAT).

  https://prometheus.io/docs/instrumenting/writing_exporters/#types

In practice, Prometheus treats untyped the same as gauge. However, other
systems ingesting these metrics might not do the same (some might not
like untyped).

For our documentation, it has also been decided to not document as
untyped, but as gauge instead, which we have been doing
(metrics_metadata.json reports as gauge).

This patch defaults metrics to gauge, so there are no untyped metrics
returned from the Prometheus endpoint.

Change-Id: I50f7db12592b56679223247ac17ca061252ad5c5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219512
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>