MB-62775: Revert intialization of bfilter in EPVB ctor
This reverts a regression caused by refactoring done to MB-48834 where
we moved initializing the bloomfilter to EPVB ctor.

What was the regression?

This had the unintended consequence of initializing the bloomfilter when
a vbucket is created during warmup. Since this bloomfilter is not
persisted to disk & we end up creating an empty bloomfilter & we incorrectly
return a key doesn't exist (despite the item being present on disk).

When compaction eventually runs we'll update the initializing the
bloomfilter & accordingly update it with all the keys found on disk.

Change-Id: If281cd0ffa793992fe75dfff042cfb5cbb3e9045
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219535
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>