[BP] MB-63996: Close access log when write to disk fails
When the MutationLog.blockBuffer is almost full & a write to disk
fails, we incorrectly increment the blockPos (position in the block
buffer where a mutation log entry is written at) beyond the
blockBuffer size.

When the disk write fails - throw an exception & disable the log.

Eventually when the item access visitor completes, skip replacing the
old access log with what is now a potentially corrupt/incomplete new
access log.

Change-Id: I6db0c3f66cf74783f7a739192f99217b0c4bc850
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/221604
Well-Formed: Restriction Checker
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>