MB-64293: Remove failing unit test for timestamp compatability
This test is failing on neo, the test has been removed on Trinity
and Master. Currently, when we check if a timestamp is compatible
with PST, we check if “-07:00“ is at the end of the timestamp.
However, when daylight hours change this difference becomes
“08:00“.

We have migrated from ISOTime to platform cb::time::timestamp
(MB-53572) and these tests have been removed in line with Master
and Trinity.

Change-Id: I5d3f12d167e05df10ce06c2f226e5b8e23723719
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219700
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>