MB-48834: Initialize kvStoreBfilterEnabled in EPVB ctor
This bool indicates if we should consult the bloomfilter maintained by
the storage backend. Magma is the only storage backend which maintains
it's own bloomfilter.

The EPVB ctor is called during warmup & when set_vbucket_state commands
are performed.

The magma bloomfilter is saved to disk unlike the one kv_engine
maintains for couchstore. Therefore for Magma, the gates to the bloomfilter can be opened even when a vbucket is constructed during warmup, unlike couchstore where we have to wait for the first compaction run to complete to correctly populate the bloomfilter based on the items present on disk.

Change-Id: I71060928f56a210b6c39230ea7f3a3dac0be95cb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219838
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>