MB-64366: Set the KVStore quota proportional to the number of vBuckets
Magma KV shards have their own quota set by
MagmaKVStore::setMaxDataSize. Currently, every shard gets 1/N of the
total Magma quota.

However, in some cases, the shards can be imbalanced. This is especially
true when the number of vBuckets comparable or lower than the number of
shards.

With this change, we give each shard a ratio "alive vBucket ratio" which
is the ratio of alive vBuckets in the shard to all alive vBuckets.

Change-Id: I9f0b73651c30cf2faf64da858b711e6ced572fa8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/219967
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>