MB-64123: Move snapshot reloading to KVStore with validation
Move I/O code from Cache to KVStore and validate snapshots before
adding to cache.

An invalid snapshot is one which is:

* missing the manifest.json - prepare interrupted
* has a sha512 mismatch - data corruption

Valid snapshots are then also checked for a number of conditions:

* An absent file - download interrupted before reaching that file
* A truncated file - download interrupted before finishing that file

A snapshot with one or more file in this condition could be resumed
if the source still holds a matching snapshot (matching by uuid).

These states are reflected in the Manifest and currently visible via
stats.

The sha512 check may be removed or kept as we have to read back all
snapshots on creating a bucket.

Change-Id: Ibab9db959645fa20c4cf7da2e0640256797ba540
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220162
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>