MB-64391: Use stack allocated buffer for the key in doEviction
When we evict a key from the HT, we create a heap allocated copy of the
key as a StoredDocKey. This shows up when profiling as taking ~40% of
the CPU time spent within the ItemPager visitor.

We can optimise this by stack-allocating the key. When we do, we see a
smaller improvement in the single-threaded ItemPager benchmarks,
although I suspect we’re getting better use of the tcache in those than
we would in production.

Change-Id: Ic19aabb37596100c86ba138a61ee76024c2d9c08
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220038
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>