MB-62703: Slow down frequency of asking for disk space usage.
For the slow DCP backfill detection path avoid reading the disk space
info or asking magma for how much maybe free on each invocation of the
check.

This is done by wrapping those values in an AutoRefreshedValue class
that provides a stale value which is refreshed when a define duration
passes.

Both values have a fixed (not dynamic) 10 second refresh duration.

Change-Id: I7bbd8f18ef129da3a21b9b398e215149088c4891
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220652
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>