MB-64340: Add tracing support to EpTask & VBCBAdaptor
Added cb::executor::Tracer which enabled tracing support in the
executor. It can be used with the existing tracing helpers like
ScopeStopwatch, etc. The EventId type it uses is EventLiteral - a
wrapper around a string literal identifying an event.

The cumulative execution time associated with a particular event is then
printed in the Slow runtime log messages.

Example (assuming we have vbcb_adaptor.visit_bucket):

```
  <TS> Slow runtime {
    "bucket": "SynchronousEPEngine:default",
    "task": "Expired item remover on vb:0",
    "thread": "checked_executor",
    "runtime": {"ms": 101.585125},
    "profile": {
      "vbcb_adaptor.visit_bucket": [48,{"ms":101.576}]
    }
  }
```

The cb::executor::Tracer is a thread_local object accessible by engine
tasks. It accumulates runtimes in a F14VectorMap<EventLiteral,
Duration>. F14VectorMap is used because it doesn't free up memory when
clear() is called. Therefore, once populated, we don't need to keep
reallocating the map as part of every task execution.

This patch also adds support for vBucket visitors to access the Tracer,
such that execution time can be reported.

Change-Id: I7b0cefcb97b5d2f433de0392ae36438b975b906a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220398
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>