MB-64353: Maintain a rollback point for ephemeral failover
getPersistedSnapshot currently returns {0,0} for ephemeral as it lacks
code to maintain the snapshot (persistent uses the flusher). This
means that hard failover will always insert a failover branch point
with seqno 0.

For ephemeral (and in later persistent) we can maintain the correct
failover branch point inside the CheckpointManager.

CheckpointManager::lastSnapshotHighSeqno tracks the highest consistent
seqno which should be used for failover entry.

This value begins as 0 and as a CM receives mutations the lastBySeqno
and current snasphot.end are compared. Once CM observes that the
lastBySeqno is equal to the snapshot.end lastSnapshotHighSeqno is
assigned the snapshot.end value.

[a..........b]  [c..........d]
            ^         ^
            X         Y

 X = lastSnapshotHighSeqno
 Y = lastBySeqno

Thus a failover will use X (b) as the branch point.

a..........b]  [c..........d]
                           ^
                           XY

 X = lastSnapshotHighSeqno
 Y = lastBySeqno

Now on receive of seqno d lastSnapshotHighSeqno moves to d.

Change-Id: Ide10b57064c03dd7c37b79c91e846f4747f7f055
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220596
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker