Fix race in ExecutorPoolEpEngineTest
The test checks the bucket memory by calling getPreciseTotalMemoryUsed
when notified. The notification happens when a test runs and calls
notify_one from it's run() method.

This is racy because after run(), we step out into GlobalTask::execute()
which does additional work, including checking if the task took too long
and logging. If anything in execute() allocates memory, the test's
getPreciseTotalMemoryUsed will race with that allocation and the test
will fail.

This can be triggered by the tracing changes and is hard to hit locally
but very easy to hit in CV. The allocation in question there appears to
be the allocation for the task name string.

Change-Id: I42dc21c93d82df97226d98741bd2c7f5520c046b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220807
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>