Fix intermittent failure in ContinousBackupTest.PathConfig
The test was supposed to verify that a path with relative elements
becomes converted to a canonical path. Since the path started with ../,
the resulting path was outside of the test's local directory, and all
tests shared the same continuous backup folder.

Perhaps avoiding that would fix the race seen on Windows.

Change-Id: I308c9dd8c24795a6a928761ab498282a9928c0d3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/220839
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>