MB-64728: Disable the WorkLoadMonitor
Added a configuration key workload_monitor_enabled and disabled by
default.

The WorkLoadMonitor was previously non-functional in 7.1.x and 7.2.x
due to a bug which has since been resolved, although it wasn't
explcitly disabled.

We've seen some regressions in 7.6.x caused by the WorkLoadMonitor, so
here we are disabling it by default which effectively puts us on par
with 7.1.x and 7.2.x.

Change-Id: Idbf771c72bab577d1f85edb14bfbeff1af8dafc3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/221400
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>