[BP] MB-64729: Fix unit tests using 01.01.2025 as exp time
Some of the old unit tests testing "return meta" command used
Jan 1st 2025 as the expiry time. If the provided expiry time
is in the past we reset it to 1 sec expiry to the current
time causing the unit tests to fail as it tested that the
provided expiry time was set.

Fixed the test to set the expiry time to 1 hour from the
current time.

Change-Id: Ic706ec9738956a21c18a78046cdf35a746216be6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223216
Tested-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker