MB-51689: Include HPS in Snapshot marker v2.2
Currently, when a replica receives a disk snapshot we set the high
prepare seqno to the snap-end, since the prepares on disk might have
been completed & the snapshot might not receive the prepares to derive
actual high prepare seqno.

Pack the high prepared seqno present in the vbstate on disk in the
Snapshot Marker v2.2. The replica will use the high prepared seqno
present on the checkpoint when it persists the vbstate to disk.

Change-Id: Icbf6c068f6a66c4b34742b689760ca536422f77e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216147
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>