MB-64825: Add back bucket level privileges for no bucket
We don't really want all clients to get EACCESS returned when
they're bound to "no bucket" (which happens before we even
try to call down in the engine). Instead we want these clients
to be allowed to call down into the engine and have the
engine return "no bucket" in the engine API which would then
go back to the client.

There is however a subset of STAT commands which are
"privileged" and allowed to be executed towards "no bucket"
(for instance listing connections, scheduler times, etc).
All of these commands needs to be able to get pass the
global privilege check for the bucket in order to allow
for execution.

We correctly added the bucket privilege when the connection
was assigned to the "no bucket" instance, but whenever the
privilege context got stale (due to modifications to the RBAC
data) all bucket privileges for "no bucket" got lost causing
all operations to return EACCESS.

This patch adds back the bucket level privileges to connections
bound to no bucket as part of refreshing the privilege context
restoring the behavior of letting the underlying engine
determine the result of the operation.

Change-Id: I3eb7a1982bc068569570891289723ebd6715dbf8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/221808
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>