MB-64849: Refactor Warmup members to use a single vector
To progress towards a fix for MB-64849 there are a number of
Warmup members which need better locking. Code review notes
that indvidual data structures exist with their own locks or
no lock, including some redundant data structures.

With refactoring we can remove

shardVbStates

This vector holds a vector per shard. The inner vector is all the
vbucket_state objects for each VBucket we are warming up.

warmedUpVbuckets

This map holds shared_ptr of the VBuckets but only between
CreateVBucket and PopulateVbucket. This map is not sharded, but
the shared_ptr can just be moved into the sharded vector.

weakVbMap

This map holds a weak_ptr of the VBuckets throughout all Warmup.
This map is not sharded, but the shared_ptr can just be moved into
the sharded vector.

Now one sharded vector owns the vbucket_state, shared_ptr and weak_ptr.

Change-Id: Ic35ca728f20d05c69d1816a21c246d7b53465191
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222012
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>