MB-64849: Refactor various Warmup members under a single mutex
This commit moves various member data which has their own locks and
some which do not (some of which are certainly accessed from different
threads like corruptAccessLog).

There is now a unified struct managed by folly::Synchronized that
stores various Warmup variables and containers.

Many other members remain unchanged as they are explicily atomic
(which may or may not be for performance reasons, but for now is left
untouched).

There should be no functional change with this commit but now provides
a way to safely change the doneFunction concurrently whilst warming up.

Change-Id: I887aa1ed585930c596476397600832c71e69cf60
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222067
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>