MB-64849: Fix Warmup stop issues
The stop can occur at anytime, whilst warmup is in any phase.
If the done callback, EPBucket::primaryWarmupCompleted is
executed from the shutdown thread.

1) A data-race can occur (which created the MB)
2) Invalid memory access can occur (as creating secondary warmup
   leaves the primary warmup object in a non runnable state).

The main fix for this is to set the done callback todo nothing.
A second issue was noted in that secondary warmup is not stopped
on shutdown.

Change-Id: Ia194ff8e6aa9dc1b97c7cd7d84de59678bd961a8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222069
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>