MB-64916: Change RangeScanContinue slow SLA to 5s
RangeScanContinue is a non-blocking I/O bound command which
is also configured by the client (how much to scan). The
default 500ms is too low and for now move to 5s (which is
not a scientifically picked number).

Overall the challenge here is that a single continue could
be invoked to read the entire database, without a dynamic
SLA based on the scan size (which could be estimated) this
larger static value may still have false positives, but is
an improvement that results in less log spam.

Change-Id: Icc4cfcc3530ae7e3f18f3bf909a58adae412710b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222121
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>