[unmerged-commits] MP branches should be merged into the GA branch
Previously, we have the MP branches linked, such that 7.1.4-MP1 would be
expected to be merged into 7.1.5-MP2. That is not what we actually want
to do as MPs are supposed to only be GA + specific backports needed for
the build.

Also added missing MP branches that have been created since.

Change-Id: Ie7e5b90008826a186d76df22a638b9d69c243d85
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222200
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>