MB-64937: Only remove deleted items if warmup is done loading
Warmup relies on a CAS check to verify that the version of the document
being inserted into the HT is the correct version.

With secondary warmup, we will allow operations which, under normal
circumstances, result in a new version of the document which can be
removed from the HT (because of DELETE or TTL or SET+expiry). Then, that
CAS check mechanism does not work (since we don't have the old key's CAS
to check), so there is a danger we might insert a stale version of the
document in the HT.

To avoid this, we now keep delete metadata in the HT after persistence
(before it was removed after persistence). In a future patch, we will
forbid metadata from being evicted during warmup.

Change-Id: I02830c0531a323a4d93b04acaaabb13abfd2222b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222520
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>