MB-64931: End Stream when start seqno > end seqno
As a part of MB-40383, we added an optimization where we fudge the start
seqno for collection filtered streams to point to the start seqno of a
collection during backfill to avoid reading mutations not relevant to a
collection.

But that could mean the adjusted start seqno is greater than the the
end seqno requested by the DCP client. End the stream if that happens.

Change-Id: I6a122f055a78326fe296f3b2db60af6401499d21
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222344
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>