MB-64937: Allow getSecondaryWarmup() without a lock
The secondary warmup is initialised after primary is done, and that is
done under a lock. The way the pointer is stored, we cannot access it
without locking the same mutex.

To allow lock-free access, we can store a copy of the raw pointer behind
an std::atomic<Warmup*>, once the secondary warmup object has been
constructed - past that point, we never reset the pointer anyways.

Change-Id: I2a916720af0793a03ce30b381944cdab69b4d9dc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222830
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>