MB-64246: Fix CheckpointManager::getSnapshotInfo()
Recent changes on ItemExpel (MB-63341) interfere in a bad way with the
logic CheckpointManager::getSnapshotInfo().

By that, if a CheckpointManager:
 - has an open checkpoint that contains only meta-items
 - and, that checkpoint is touched by ItemExpel
then the CM might expose invalid snapshot information by
CheckpointManager::getSnapshotInfo().

That API is mainly used by DCP streams for determining the in-memory
snapshot state of the related vbucket.
In the specific case of MB-64246 a PassiveStream ends up making a
invalid StreamRequest to the Producer. The Producer validates the
StreamRequest and fails the request. PassiveStream enters a loop by
issuing the same invalid StreamRequest over and over.

In this patch CheckpointManager::getSnapshotInfo() is fixed to handle
the MB-63341 case and its logic is also simplified. Details in
comments.

Change-Id: I862a3ffd6d16fe1199df5ac99e60e4283d2049e0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222523
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>