MB-64896: Fix flusher_todo stat
The flusher_todo stat was previously reset to 0 on a completion of a
flush batch. This could cause an issue where one flusher increments the
stat while another flusher completes and sets it to 0, losing the
increment.

The flusher_todo stat now is incremented for each item that must be
written and is decremeneted by the same amount on the completion of each
flusher run.

Change-Id: Ica3619d2dffb15c819149300504253b54f72d445
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222768
Tested-by: Faizan Alam <faizan.alam@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>