MB-64001: Change DcpProducer to use an abstract ProducerStream
Add ProducerStream to be the parent class of ActiveStream in
readiness for the addition of CacheTransferStream's into a
producer.

In this commit the only sub-class of ProducerStream is an ActiveStream
which sources data from checkpoint or backfill.

Later in MB-64001 CacheTransferStream will be managed by a DcpProducer
and will be a sub-class of ProducerStream and source data from hash
table visiting.

Change-Id: I43fbd54967a0384ac0e97e78ceeccce343cb55f8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225699
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>