MB-64001: Add DCP CacheTransferStream/Task/Visitor
Add the following new classes.

* CacheTransferStream
* CacheTransferHashTableVisitor
* CacheTransferTask

CacheTransferStream inherits from Stream, the new stream class uses
the CacheTransferTask to visit a vbucket's hash-table and copy
eligible resident and alive values into the Stream::readyQ.

The main criteria is the "maxSeqno" which is a value that would
be provided by a replica VB that wish to copy an active VB's cache.
The replica would normally use its high-seqno in the copy and this
ensures that only the resident items which already exist in a replica
are copied, items above the maxSeqno would be moved by a normal
ActiveStream.

The visiting pattern respects the bucket's memory usage, ensuring it
will yield if mem_used is above the high-water mark (so we cannot
grow memory by creating output messages). However resuming a paused
visit has no clever wakeup mechanism other than a time-based snooze and
retry (this is deliberate as this patch intends to add the overall
stream/task/visitor infrastructure).

In this first commit there is no advanced filtering of the resident
items which may come later - e.g. "send the 10% hottest items".

Change-Id: Id3411baa3b22e717612c484175f80e4b423e9617
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227232
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>