MB-65122: non-const Stream members require streamMutex
getEndSeqno/setEndSeqno is one example where no mutex is used
yet they are reading/writing end_seqno_ which should be accessed
with the streamMutex.

getSnapEndSeqno is used in test code, and now use the mutex.

name_ is now made const.

PassiveStream::reconnectStream leaves the read (maybe it's deliberately
that way) of the latest uuid outside of the streamMutex, but the
assignment to vb_uuid_ is now under the streamMutex locked scope.

Change-Id: I01d8eda743c2a7a3386f15874e783328491b5394
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222783
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>