MB-64123: Improve snapshot download
Add an implementation in MemcachedConnection which utilize
a dedicated "read callback" which allows for writing data
to disk without buffering the entire MCBP frame in memory.
This solves the latency problem over TLS which was caused
by the client requesting small fragments of the file at
a time.

The new implementation allows for fetching up to 2GB of
data per GetFrameRequest and the receiving side drains
the socket and writes to disk.

The client (ns_server) is also able to specify the
interval between each call to fsync() to avoid having
too large segments of uncommitted data.

Change-Id: I26e659ba3fd4a7d1115294b65bd9048114c0cf8f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/222827
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>