MB-64465: Ensure Flusher cannot go from stopping to running
There is a race in how the Flusher _state is updated when a force
shutdown is being performed. This has been seen to hang tests in
CV but could also deadlock a real server shutdown.

Prior to this commit it was possible for the step function to read _state
and execute as if "Initializing", yet the shutdown can set _state to
"Stopping". As step moves to "Running" the validTransition function is
skipped because of the force shutdown logic.

This change.

- Removes the force shutdown skip of validTransition, all transitions
  must be checked.
- Add a cmpxchg loop into the transition function so that the update
  is consistent with the validated from/to.

Change-Id: I7186fe4ba205cf8312a9baf2446a7bcfa81452d5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223004
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>