MB-62963: Ephemeral buckets backfill optimization
If the start seqno is greater than the high seqno of the list of
collections requested in a collection filtered stream, we can avoid a
backfill.

This was already done for persisted buckets via MB-52957; apply the same
optimization for ephemeral buckets.

Cleanup a bunch of tests, which were skipped for ephemeral due to the
difference in behavior.

Change-Id: Ibea934e27788d42571b5fe99b04d23ba0f3d0b19
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223043
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>