MB-65019: Prevent rollback if start seqno is collection's start seqno
In MB-40383, for collections filtered streams, we set the startSeqno to
be the collection's startSeqno if it is less than it. This is to avoid
reading non-relevant items. However, if the requested startSeqno is 0,
we can modify it to be greater than 0 but still < purgeSeqno, in which
case we incorrectly request a rollback.

Check if the startSeqno == collection's startSeqno, in which case do not
request a rollback since all required items are requested.

Change-Id: I33f91debded751756bceb713f826262cb8dbb469
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223188
Tested-by: Faizan Alam <faizan.alam@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>