MB-65240: Fix disconnect for slow stats
Before this patch the DCP stat worked by first scheduling a task to
call down into the engine and spool up to 16k of stats response then
it would connect back to the front end thread which would chain the
response to the output queue with a custom deleter which would notify
the cookie (and leave the cookie in blocked state).  Once notified the
cookie would repeat the sequence again until all stats was
returned. This works perfectly well in the "happy" path, but if a
network error occured while transferring data the cookie's state would
be EWB, and when the error callback from bufferevent arrives shutdown
is blocked due to the state of the cookie. Bufferevent would *not*
drain the queue and call the destructors (causing the notification)
until the entire bufferevent object would be released.

This patch introdue a new error state too_much_data_in_output_buffer
which is "similar" to would_block, except that it does not require a
cookie notification to resume. It relies on the fact that libevent
will notify the instance once all data has been transferred into the
kernel.

Change-Id: I729a2701c4597b8ea0dde30fe53d46f6227089ac
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223143
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>